CreateModulesSourceCountMetricUseCase.kt refactored
